add-interactive.c: implement list_modified
Implement list_modified from Perl, which will be used
by most of the *_cmd functions, including status in the
following commit.

It lists a numstat comparing changed files between a) the worktree and
the index; b) the index and the HEAD.

To do so, we use run_diff_index() and run_diff_files() to get changed
files, use the diffstat API on them to get the numstat and use a
combination of a hashmap and qsort() to print the result in
O(n) + O(n lg n) complexity.

Add new file: add-interactive.c which will be used for implementing
"application logic" of git add -i (alongside with add-interactive.h,
added in later commit), whereas add--helper.c will be used mostly
for parsing the command line.

Mentored-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Original-patch-by: Daniel Ferreira <bnmvco@gmail.com>
Signed-off-by: Slavica Djukic <slawica92@hotmail.com>