add-interactive.c: implement status command
Implement add --interactive's status command in add-interactive.c and
use it in builtin add--helper.c.
This is the first interactive add command implemented in C of those
anticipated by the previous commit, which introduced
the add--helper built-in.

Implement additional helper functions dealing with struct choice and
colors.

Use list_modified to get the data, and then pass it to list_and_choose
to show it in appropriate format.

We're a bit lax with the command-line parsing, as the command is
intended to be called only by one internal user: the add--interactive script.

Mentored-by: Johannes Schindelin <Johannes.Schindelin@gmx.de>
Original-patch-by: Daniel Ferreira <bnmvco@gmail.com>
Signed-off-by: Slavica Djukic <slawica92@hotmail.com>