read-cache: add post-index-changed hook
Add a post-index-changed hook that is invoked after the index is
written in do_write_locked_index().

This hook is meant primarily for notification, and cannot affect
the outcome of git commands that trigger the index write.

The hook is passed a flag to indicate whether the working directory
was updated or not and a flag indicating if a skip-worktree bit
could have changed.  These flags enable the hook to optimize its
response to the index changed notification.

Signed-off-by: Ben Peart <benpeart@microsoft.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>