sha1-array: implement oid_array_readonly_contains
Implement a "readonly_contains" function for oid_array that won't
sort the array if it is unsorted. This can be used to test containment in
the rare situations where the array order matters.

The function has intentionally been given a name that is more cumbersome
than the "lookup" function, which is what most callers will will want
in most situations.

Signed-off-by: Stefan Xenos <sxenos@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>