archive: use the internal zlib-based gzip compression by default
We just introduced support for compressing `.tar.gz` archives in the
`git archive` process itself, using zlib directly instead of spawning
`gzip`.

While this takes less CPU time overall, on multi-core machines, this is
slightly slower in terms of wall clock time (it seems to be in the
ballpark of 15%).

It does reduce the number of dependencies by one, though, which makes it
desirable to turn that mode on by default.

Changing the default benefits most notably the MinGit flavor of Git for
Windows (which intends to support 3rd-party applications that want to
use Git and want to bundle a minimal set of files for that purpose, i.e.
stripping out all non-essential files such as interactive commands,
Perl, and yes, also `gzip`).

We also can now remove the `GZIP` prerequisite from quite a number of
test cases in `t/t5000-tar-tree.sh`.

This closes https://github.com/git-for-windows/git/issues/1970

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>