p4 unshelve: fix "Not a valid object name HEAD0"
git p4 unshelve was failing with these errors on Windows:

fatal: Not a valid object name HEAD0
Command failed: git cat-file commit HEAD^0

(git version 2.21.0.windows.1, python 2.7.16)

The pOpen call used by git-p4 to invoke the git command can take either a
string or an array as a first argument.  The array form is preferred
however the extractLogMessageFromGitCommit method was using the string
form, which makes the caller responsible for escaping the command text
appropriately (see https://docs.python.org/2/library/subprocess.html)

Somewhat ironically, the carat character is the escape character in
Windows and so should be escaped (HEAD^^0).  Without the extra carat, the
OS was passing an escaped 0 to the git command instead, and the git
command was rejecting the invalid object name "HEAD0"

The behaviour can be confirmed by typing ECHO HEAD^0 at the command-
prompt, which emits HEAD0.

The solution is simply to use the array format of passing the command to
fOpen, which is preferred and used in other parts of this code anyway.

Signed-off-by: Mike Mueller <mike.mueller@moodys.com>