bundle verify: error out if called without an object database
Bundles really are thin packs (i.e. in general, they require an object
database to reconstruct deltified objects), with very little sugar on
top. So we really need a repository (or more appropriately, an object
database) to work with, when asked to verify a bundle.

Let's error out with a useful error message if `git bundle verify` is
called without such an object database to work with.

Reported by Konstantin Ryabitsev.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>