refactor: rename is_directory() to dir_exists() and use it in clone.c
The original is_directory() checks whether the given path exists as
a directory, which makes dir_exists() a more suitable name.
However, there is already an existing function called dir_exists(),
while it doesn't check if the path is a directory.

We decided to do the following:
- remove the original dir_exists()
- rename the original is_directory() to dir_exists()
- use the new dir_exists() where the original dir_exists() is called

Hope it can reduce some confusion.

Signed-off-by: John Lin <johnlinp@gmail.com>