add--interactive: skip index refresh in reset patch mode
Uses the FILTER flag on patch flavours in git-add--interactive to
determine if the initial index refresh call is required. It is not
required for the reset patch modes, which use the 'index-only' filter,
as they do not interact with the worktree, so do not need up-to-date
stat information.

The refresh call can be quite expensive, especially on larger
checkouts or with a freshly created index file.

Signed-off-by: Nika Layzell <nika@thelayzells.com>