t3422: mark two test cases as requiring support for `--preserve-merges`
We started marking all of those test cases that require `git rebase -p`
to work in the recent past, to allow for skipping them (because the
`--preserve-merges` backend is already deprecated and will be removed in
one of the next major versions, so it is a bit pointless to test its
functionality all the time).

This patch marks two previously forgotten test cases with the `REBASE_P`
prerequisite, to allow skipping them via setting the environment
variable `GIT_TEST_SKIP_REBASE_P`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>